# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "description": "Ruta para testear el punto inicial",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inicio"
                ],
                "summary": "Ruta inicial",
                "responses": {
                    "200": {
                        "description": "API GO LANG",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ConsultarDeuda": {
            "post": {
                "description": "Obtener el listado de deudas del alumno con su código o dni",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consultar"
                ],
                "summary": "Obtener lista de deudas del alumno",
                "parameters": [
                    {
                        "description": "Estructura para realizar la consulta",
                        "name": "opcion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BodyConsultarDeuda"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BodyConsultarDeudaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/NotificarPago": {
            "post": {
                "description": "Proceso para realizar el pago correspondientes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notificar Pago"
                ],
                "summary": "Registrar pago",
                "parameters": [
                    {
                        "description": "Estructura para realizar la consulta",
                        "name": "opcion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BodyNotificarPago"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BodyNotificarPagoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.BodyConsultarDeuda": {
            "type": "object",
            "properties": {
                "ConsultarDeuda": {
                    "$ref": "#/definitions/model.ConsultarDeuda"
                }
            }
        },
        "model.BodyConsultarDeudaResponse": {
            "type": "object",
            "properties": {
                "ConsultarDeudaResponse": {
                    "$ref": "#/definitions/model.ConsultarDeudaResponse"
                }
            }
        },
        "model.BodyNotificarPago": {
            "type": "object",
            "properties": {
                "NotificarPago": {
                    "$ref": "#/definitions/model.NotificarPago"
                }
            }
        },
        "model.BodyNotificarPagoResponse": {
            "type": "object",
            "properties": {
                "NotificarPagoResponse": {
                    "$ref": "#/definitions/model.NotificarPagoResponse"
                }
            }
        },
        "model.Cabecera": {
            "type": "object",
            "properties": {
                "operacion": {
                    "$ref": "#/definitions/model.Operacion"
                }
            }
        },
        "model.ConsultarDeuda": {
            "type": "object",
            "properties": {
                "recaudosRq": {
                    "$ref": "#/definitions/model.RecaudosRq"
                }
            }
        },
        "model.ConsultarDeudaResponse": {
            "type": "object",
            "properties": {
                "recaudosRs": {
                    "$ref": "#/definitions/model.RecaudosRs"
                }
            }
        },
        "model.Detalle": {
            "type": "object",
            "properties": {
                "respuesta": {
                    "$ref": "#/definitions/model.Respuesta"
                },
                "transaccion": {
                    "$ref": "#/definitions/model.Transaccion"
                }
            }
        },
        "model.Documento": {
            "type": "object",
            "properties": {
                "cantidadSubconceptos": {
                    "type": "integer"
                },
                "descripcion": {
                    "type": "string"
                },
                "fechaEmision": {
                    "type": "string"
                },
                "fechaVencimiento": {
                    "type": "string"
                },
                "importeDeuda": {
                    "type": "number"
                },
                "importeDeudaMinima": {
                    "type": "number"
                },
                "indicadorRestriccPago": {
                    "type": "integer"
                },
                "numero": {
                    "type": "string"
                }
            }
        },
        "model.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.NotificarPago": {
            "type": "object",
            "properties": {
                "recaudosRq": {
                    "$ref": "#/definitions/model.RecaudosRq"
                }
            }
        },
        "model.NotificarPagoResponse": {
            "type": "object",
            "properties": {
                "recaudosRs": {
                    "$ref": "#/definitions/model.RecaudosRs"
                }
            }
        },
        "model.Operacion": {
            "type": "object",
            "properties": {
                "canalOperacion": {
                    "type": "string"
                },
                "codigoBanco": {
                    "type": "integer"
                },
                "codigoConvenio": {
                    "type": "integer"
                },
                "codigoOficina": {
                    "type": "string"
                },
                "codigoOperacion": {
                    "type": "integer"
                },
                "fechaOperacion": {
                    "type": "string"
                },
                "horaOperacion": {
                    "type": "string"
                },
                "numeracionOperacion": {
                    "type": "integer"
                }
            }
        },
        "model.RecaudosRq": {
            "type": "object",
            "properties": {
                "cabecera": {
                    "$ref": "#/definitions/model.Cabecera"
                },
                "detalle": {
                    "$ref": "#/definitions/model.Detalle"
                }
            }
        },
        "model.RecaudosRs": {
            "type": "object",
            "properties": {
                "cabecera": {
                    "$ref": "#/definitions/model.Cabecera"
                },
                "detalle": {
                    "$ref": "#/definitions/model.Detalle"
                }
            }
        },
        "model.Respuesta": {
            "type": "object",
            "properties": {
                "codigo": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                }
            }
        },
        "model.Transaccion": {
            "type": "object",
            "required": [
                "numeroReferenciaDeuda"
            ],
            "properties": {
                "cantidadDocsDeuda": {
                    "type": "integer"
                },
                "codigoMoneda": {
                    "type": "string"
                },
                "datosEmpresa": {
                    "type": "string"
                },
                "fechaOperacionOriginal": {
                    "type": "string"
                },
                "formaPago": {
                    "type": "string"
                },
                "importeDeudaPagada": {
                    "type": "number"
                },
                "indMasDeuda": {
                    "type": "integer"
                },
                "listaDocumentos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Documento"
                    }
                },
                "nombreCliente": {
                    "type": "string"
                },
                "numeroDocumento": {
                    "type": "string"
                },
                "numeroOperacionEmpresa": {
                    "type": "integer"
                },
                "numeroOperacionOriginal": {
                    "type": "integer"
                },
                "numeroOperacionRecaudos": {
                    "type": "integer"
                },
                "numeroReferenciaDeuda": {
                    "type": "string"
                },
                "otrosDatosEmpresa": {
                    "type": "string"
                }
            }
        }
    }
}